Revise tests to check forms directly
Testing `IsSubset(G, elm)` can be slow in GAP if G is a matrix group of
"large" degree. This is because GAP resorts to testing membership using a
stabilizer chain in an isomorphic permutation group, even if G is a
linear/unitary/orthogonal/symplectic group. So we replace that check with one
that actually tests using the invariant forms of the groups in question.