Speed up GammaLMeetSU and UnitarySemilinearSp
... by avoiding LogFFE, achieved by creating the generators for SU resp. GU
"from scratch" (with code adapted from the GAP library), but with a twist:
instead of finite field elements, we use rational functions that express the
entries in terms of a primitive element. This way the map Theta (implemented
here via `MapGammaLToGLRatFun` as a variation of `MapGammaLToGL`) is easy to
compute: we just evaluate each non-zero rational function in the entries of
the generator matrices at the matrix "A"

Before this patch:

    gap> GammaLMeetSU(9, 8, 3);; time;
    3396
    gap> UnitarySemilinearSp(8,3^6);; time;
    8860

After this patch:

    gap> GammaLMeetSU(9, 8, 3);; time;
    19
    gap> UnitarySemilinearSp(8,3^6);; time;
    135

Also simplify `GUMinusSU` -- while not needed for this change (we replace
a call to `GUMinusSU`), it is useful in that it makes it obvious that
the replacement code for it corresponds to what `GUMinusSU` does.

Finally, replace two problematic uses of `PrimitiveElement` by
`PrimitiveRoot`: they were problematic because they were passed to
`MapGammaLToGL` which in turn passes them to LogFFE; which by definition only
works in general for primitive roots.