Remove hacks that allowed passing GF(q) instead of GF(q^2)
These hacks were needed back when we did not pass around the fields
explicitly. But they always were a crutch: they help us to deal with
the case where e.g. q = p or q = p^3 ("so the user already took the
square root")

Some tests relied on the removed workaround, so adjust those as well.